Move print and equal to their own lib files
