+++ runtime/src/lib/equal.js
@@ -1,0 +1,40 @@
+import { GRAIN_TRUE, GRAIN_FALSE } from '../core/primitives';
+import { view } from '../runtime';
+
+function equalHelp(x, y, cycles) {
+  if ((x & 7) === 1) {
+    if ((y & 7) !== 1) {
+      return false;
+    }
+    let xPtr = (x ^ 1) / 4;
+    let yPtr = (y ^ 1) / 4;
+    if (view[xPtr] !== view[yPtr]) {
+      return false;
+    }
+    if (view[xPtr] & 0x80000000) {
+      return true;
+    }
+    let length = view[xPtr];
+    ++cycles;
+    view[xPtr] |= 0x80000000;
+    view[yPtr] |= 0x80000000;
+    let result = true;
+    for (let i = 0; i < length; ++i) {
+      if (!equalHelp(view[xPtr + i + 1],
+                          view[yPtr + i + 1],
+                          cycles)) {
+        result = false;
+        break;
+      }
+    }
+    view[xPtr] = length;
+    view[yPtr] = length;
+    return result;
+  } else {
+    return x === y;
+  }
+}
+
+export default function equal(x, y) {
+  return equalHelp(x, y, 0) ? GRAIN_TRUE : GRAIN_FALSE;
+}
