+++ runtime/src/lib/print.js
@@ -1,6 +1,11 @@
 import { grainToString } from '../utils/utils';
 
-export default function print(v) {
+export function print(v) {
   console.log(grainToString(v));
+  return v;
+}
+
+export function debugPrint(v) {
+  console.log(`0x${v.toString(16)} (0b${v.toString(2)})`);
   return v;
 }
